requirements.txt
openai
anthropic
python-pptx>=0.6.21
python-dotenv
pyyaml
aiohttp
pytest
Pillow>=10.0.0
